Overhaul auto-config
