--- src/main/java/io/clue2solve/aws/bedrock/springboot/starter/autoconfigure/clients/BedrockClientConfig.java
@@ -1,89 +1,0 @@
-package io.clue2solve.aws.bedrock.springboot.starter.autoconfigure.clients;
-
-import org.springframework.beans.factory.annotation.Value;
-import org.springframework.boot.autoconfigure.condition.ConditionalOnProperty;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-import software.amazon.awssdk.auth.credentials.AwsCredentialsProvider;
-import software.amazon.awssdk.regions.Region;
-import software.amazon.awssdk.services.bedrock.BedrockAsyncClient;
-import software.amazon.awssdk.services.bedrock.BedrockClient;
-import software.amazon.awssdk.services.bedrockruntime.BedrockRuntimeAsyncClient;
-import software.amazon.awssdk.services.bedrockruntime.BedrockRuntimeClient;
-
-/**
- * Configuration class for BedrockClient
- */
-@Configuration
-public class BedrockClientConfig {
-
-	/**
-	 * AwsCredentialsProvider
-	 */
-	private AwsCredentialsProvider awsCredentialsProvider;
-
-	/**
-	 * Region
-	 */
-	@Value("${aws.region}")
-	private String region;
-
-	/**
-	 * Constructor
-	 * @param awsCredentialsProvider AwsCredentialsProvider
-	 */
-	@Autowired
-	public BedrockClientConfig(AwsCredentialsProvider awsCredentialsProvider) {
-		this.awsCredentialsProvider = awsCredentialsProvider;
-	}
-
-	/**
-	 * Bean for BedrockClient
-	 * @return BedrockClient
-	 */
-	@Bean
-	@ConditionalOnProperty(name = "bedrock.client.type", havingValue = "sync")
-	public BedrockClient bedrockClient() {
-		return BedrockClient.builder().region(Region.of(region)).credentialsProvider(awsCredentialsProvider).build();
-	}
-
-	/**
-	 * Bean for BedrockRuntimeClient
-	 * @return BedrockRuntimeClient
-	 */
-	@Bean
-	public BedrockRuntimeClient bedrockRuntimeClient() {
-		return BedrockRuntimeClient.builder()
-			.region(Region.of(region))
-			.credentialsProvider(awsCredentialsProvider)
-			.build();
-	}
-
-	/**
-	 * Bean for BedrockRuntimeAsyncClient
-	 * @return BedrockRuntimeAsyncClient
-	 */
-	@Bean
-	@ConditionalOnProperty(name = "bedrock.client.type", havingValue = "async")
-	public BedrockRuntimeAsyncClient bedrockRuntimeAsyncClient() {
-		return BedrockRuntimeAsyncClient.builder()
-			.region(Region.of(region))
-			.credentialsProvider(awsCredentialsProvider)
-			.build();
-	}
-
-	/**
-	 * Bean for BedrockAsyncClient
-	 * @return BedrockAsyncClient
-	 */
-	@Bean
-	@ConditionalOnProperty(name = "bedrock.client.type", havingValue = "async")
-	public BedrockAsyncClient bedrockAsyncClient() {
-		return BedrockAsyncClient.builder()
-			.region(Region.of(region))
-			.credentialsProvider(awsCredentialsProvider)
-			.build();
-	}
-
-}
