+++ src/test/java/io/clue2solve/aws/bedrock/springboot/starter/TestInit.java
@@ -1,10 +1,11 @@
 package io.clue2solve.aws.bedrock.springboot.starter;
 
+import io.clue2solve.aws.bedrock.springboot.starter.autoconfigure.aimodels.BedrockAutoConfiguration;
+import org.springframework.boot.autoconfigure.ImportAutoConfiguration;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
-import org.springframework.boot.context.properties.ConfigurationPropertiesScan;
 
 @SpringBootApplication
-@ConfigurationPropertiesScan
+@ImportAutoConfiguration(BedrockAutoConfiguration.class)
 public class TestInit {
 
 }
